# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trust Bank Marketing & Sales Suite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get the current acting user (demo user switcher)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: setCurrentUser
      tags: [users]
      summary: Switch current acting user (demo)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetCurrentUserBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - in: query
          name: stage
          schema: { type: string, enum: [new, contacted, qualified, converted, lost] }
        - in: query
          name: assignedToId
          schema: { type: integer }
        - in: query
          name: campaignId
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDetail"
    patch:
      operationId: updateLead
      tags: [leads]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      parameters:
        - in: query
          name: channel
          schema: { type: string, enum: [sms, email, in_app] }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/send:
    post:
      operationId: sendNotification
      tags: [notifications]
      summary: Send a single or bulk notification (simulated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendNotificationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendNotificationResult"
...
  /notifications/inbox/{id}/read:
    post:
      operationId: markInboxAlertRead
      tags: [notifications]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxAlert"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [marketing, sales, manager] }
        avatarColor: { type: string }
      required: [id, name, email, role, avatarColor]
    SetCurrentUserBody:
      type: object
      properties:
        userId: { type: integer }
      required: [userId]
...
    Lead:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: string }
        email: { type: string, nullable: true }
        productInterest: { type: string }
        source: { type: string }
        campaignId: { type: integer, nullable: true }
        campaignName: { type: string, nullable: true }
        assignedToId: { type: integer, nullable: true }
        assignedToName: { type: string, nullable: true }
        stage: { type: string, enum: [new, contacted, qualified, converted, lost] }
        createdAt: { type: string }
      required: [id, name, phone, productInterest, source, stage, createdAt]
    LeadDetail:
      allOf:
        - $ref: "#/components/schemas/Lead"
        - type: object
          properties:
            interactions:
              type: array
              items:
                $ref: "#/components/schemas/Interaction"
          required: [interactions]
    CreateLeadBody:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        email: { type: string, nullable: true }
        productInterest: { type: string }
        source: { type: string }
        campaignId: { type: integer, nullable: true }
        assignedToId: { type: integer, nullable: true }
      required: [name, phone, productInterest, source]
    UpdateLeadBody:
      type: object
      properties:
        stage: { type: string, enum: [new, contacted, qualified, converted, lost] }
        assignedToId: { type: integer, nullable: true }
        productInterest: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string }
        name: { type: string }
    Interaction:
      type: object
      properties:
        id: { type: integer }
        leadId: { type: integer }
        leadName: { type: string }
        type: { type: string, enum: [call, meeting, branch_visit, sms, email] }
        notes: { type: string }
        salesOfficerId: { type: integer }
        salesOfficerName: { type: string }
        nextFollowUp: { type: string, nullable: true }
        createdAt: { type: string }
      required: [id, leadId, leadName, type, notes, salesOfficerId, salesOfficerName, createdAt]
...
    Notification:
      type: object
      properties:
        id: { type: integer }
        channel: { type: string, enum: [sms, email, in_app] }
        recipient: { type: string }
        subject: { type: string, nullable: true }
        message: { type: string }
        status: { type: string, enum: [queued, sent, delivered, failed] }
        recipientCount: { type: integer }
        sentByName: { type: string }
        campaignName: { type: string, nullable: true }
        createdAt: { type: string }
      required: [id, channel, recipient, message, status, recipientCount, sentByName, createdAt]
    SendNotificationBody:
      type: object
      properties:
        channel: { type: string, enum: [sms, email, in_app] }
        audience: { type: string, enum: [all_customers, account_holders, loan_clients, business_customers, staff, custom] }
        recipients:
          type: array
          items: { type: string }
          nullable: true
        subject: { type: string, nullable: true }
        message: { type: string }
        campaignId: { type: integer, nullable: true }
      required: [channel, audience, message]
    SendNotificationResult:
      type: object
      properties:
        notification:
          $ref: "#/components/schemas/Notification"
        deliveredCount: { type: integer }
      required: [notification, deliveredCount]
    InboxAlert:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        body: { type: string }
        category: { type: string, enum: [new_lead, follow_up, campaign, deal, system] }
        read: { type: boolean }
        createdAt: { type: string }
      required: [id, title, body, category, read, createdAt]
...