# lib/api-zod/src/generated/types/leadDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trust Bank Marketing & Sales Suite API
 * OpenAPI spec version: 0.1.0
 */
import type { Interaction } from "./interaction";
import type { Lead } from "./lead";

export type LeadDetail = Lead & {
  interactions: Interaction[];
};