# lib/api-zod/src/generated/types/sendNotificationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trust Bank Marketing & Sales Suite API
 * OpenAPI spec version: 0.1.0
 */
import type { SendNotificationBodyAudience } from "./sendNotificationBodyAudience";
import type { SendNotificationBodyChannel } from "./sendNotificationBodyChannel";

export interface SendNotificationBody {
  channel: SendNotificationBodyChannel;
  audience: SendNotificationBodyAudience;
  recipients?: string[] | null;
  subject?: string | null;
  message: string;
  campaignId?: number | null;
}