# lib/api-zod/src/generated/types/setCurrentUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trust Bank Marketing & Sales Suite API
 * OpenAPI spec version: 0.1.0
 */

export interface SetCurrentUserBody {
  userId: number;
}